[step3] feat(lotto) 2등 추가  , PrizeInfromation -> Rank 로 변경
